move context logger out of util
